+++ main.py
@@ -1,11 +1,9 @@
-
 from fastapi import FastAPI, Request
 from fastapi.templating import Jinja2Templates
 
-from constants import FREQUENCY, SEX, LEVEL_OF_RESPONSIVENESS, HEART_RATE, HEART_STRENGTH, \
-HEART_RHYTHM, RESPIRATORY_RATE, RESPIRATORY_RHYTHM, RESPIRATORY_EFFORT, SKIN_COLOR, SKIN_TEMPERATURE, SKIN_MOISTURE, \
-BODY_TEMPERATURE, PUPILS, BLOOD_PRESSURE
-from models import Symptom, Condition, PatientVitals
+from models import SEX, LEVEL_OF_RESPONSIVENESS, HEART_RATE, HEART_STRENGTH, HEART_RHYTHM, RESPIRATORY_RATE, \
+RESPIRATORY_RHYTHM, RESPIRATORY_EFFORT, SKIN_COLOR, SKIN_TEMPERATURE, SKIN_MOISTURE, BODY_TEMPERATURE, PUPILS, \
+BLOOD_PRESSURE, Symptom, Condition, PatientVitals
 
 
 app = FastAPI()
@@ -44,7 +42,8 @@ def read_root(request: Request):
     print(condition)
 
 
-    return templates.TemplateResponse(
-        request=request, name="index.html", context={"patient": patient}
+    return templates.TemplateResponse(request=request, name="index.html", context={"patient": patient,
+                                                                                   "selected_symptoms": selected_symptoms,
+                                                                                   "condition": condition}
     )
