+++ templates/index.html
@@ -4,6 +4,8 @@
         <title>WFR Review</title>
     </head>
     <body>
-        {{ patient.heart_rate }}
+        {{ patient }}
+        {{ selected_symptoms }}
+        {{ condition}}
     </body>
 </html>
